backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.12.5
python-dotenv==1.0.1
reportlab==4.2.5
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
